nookLibrary: Added option to generate a screen-saver image depicting the covers of the books last accessed.
The application's main menu provides a corresponding option, which allows for defining a name for the screen-saver.  If set, a corresponding directory will be created in the "my screensavers" directory on the internal SD card.

Internally, the screen saver name is stored in a shared (String) preference named "SCREENSAVER_FOLDER_NAME".  By default, the last-accessed (up to) 5 books' covers will be rendered in the image; this number may be changed by specifying an alternate (integer) value in the shared preference "SCREENSAVER_COVER_LIMIT" (values larger than 5 will not yield any more covers being rendered for reasons of visibility and performance).